chore: make delete file in batch to parallel
