+++ src/query/storages/fuse/src/io/files.rs
@@ -12,44 +12,53 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-use std::mem;
 use std::sync::Arc;
 
 use common_catalog::table_context::TableContext;
 use common_exception::Result;
 use opendal::Operator;
 
+use crate::io::try_join_futures;
+
 // File related operations.
 pub struct Files {
+    ctx: Arc<dyn TableContext>,
     operator: Operator,
 }
 
 impl Files {
-    pub fn create(_: Arc<dyn TableContext>, operator: Operator) -> Self {
-        Self { operator }
+    pub fn create(ctx: Arc<dyn TableContext>, operator: Operator) -> Self {
+        Self { ctx, operator }
     }
 
+    /// Removes a batch of files asynchronously by splitting a list of file locations into smaller groups of size 1000,
+    /// and then deleting each group of files using the delete_files function.
     #[tracing::instrument(level = "debug", skip_all)]
     pub async fn remove_file_in_batch(
         &self,
         file_locations: impl IntoIterator<Item = impl AsRef<str>>,
     ) -> Result<()> {
-        let iter = file_locations.into_iter().map(|v| v.as_ref().to_string());
+        let batch_size = 1000;
+        let locations = Vec::from_iter(file_locations.into_iter().map(|v| v.as_ref().to_string()));
+        let mut chunks = locations.chunks(batch_size);
 
-        // OpenDAL support delete via a stream. But it will cause higher
-        // ranked problems. So let's write them by hand.
-        let mut paths = Vec::with_capacity(1000);
-        for path in iter {
-            paths.push(path);
-            if paths.len() >= 1000 {
-                self.operator.remove(mem::take(&mut paths)).await?;
-            }
-        }
+        let tasks = std::iter::from_fn(move || {
+            chunks
+                .next()
+                .map(|location| Self::delete_files(self.operator.clone(), location.to_vec()))
+        });
 
-        if !paths.is_empty() {
-            self.operator.remove(mem::take(&mut paths)).await?;
-        }
+        try_join_futures(
+            self.ctx.clone(),
+            tasks,
+            "batch-remove-files-worker".to_owned(),
+        )
+        .await?;
+        Ok(())
+    }
 
+    async fn delete_files(op: Operator, locations: Vec<String>) -> Result<()> {
+        op.remove(locations).await?;
         Ok(())
     }
 }
