+++ src/query/storages/fuse/src/io/mod.rs
@@ -14,6 +14,7 @@
 
 mod files;
 mod locations;
+mod parallel;
 mod read;
 mod segments;
 mod snapshots;
@@ -21,6 +22,8 @@ mod write;
 
 pub use files::Files;
 pub use locations::TableMetaLocationGenerator;
+pub use parallel::try_join_futures;
+pub use parallel::try_join_futures_with_vec;
 pub use read::BlockReader;
 pub use read::BloomBlockFilterReader;
 pub use read::MergeIOReadResult;
@@ -31,8 +34,6 @@ pub use read::SegmentInfoReader;
 pub use read::SnapshotHistoryReader;
 pub use read::TableSnapshotReader;
 pub use read::UncompressedBuffer;
-pub use segments::try_join_futures;
-pub use segments::try_join_futures_with_vec;
 pub use segments::SegmentsIO;
 pub use snapshots::ListSnapshotLiteOption;
 pub use snapshots::SnapshotLiteListExtended;
